requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-migrate==3.1.0
python-dotenv==0.19.0
pyjwt==2.3.0
gunicorn==20.1.0